fix: fix write partition table
